Relocated wopibridge source code
