--- .drone_bridge.yml
@@ -1,66 +1,0 @@
----
-kind: pipeline
-type: docker
-name: release-latest
-
-platform:
-  os: linux
-  arch: amd64
-
-trigger:
-  branch:
-  - master
-  event:
-    exclude:
-    - pull_request
-    - tag
-    - promote
-    - rollback
-
-steps:
-- name: publish-docker-wopibridge-latest
-  pull: always
-  image: plugins/docker
-  settings:
-    repo: cs3org/wopibridge
-    tags: latest
-    dockerfile: wopibridge-poc.Dockerfile
-    username:
-      from_secret: dockerhub_username
-    password:
-      from_secret: dockerhub_password
-    custom_dns:
-      - 128.142.17.5
-      - 128.142.16.5
-
----
-kind: pipeline
-type: docker
-name: release
-
-platform:
-  os: linux
-  arch: amd64
-
-trigger:
-  event:
-    include:
-    - tag
-
-steps:
-- name: publish-docker-wopibridge-tag
-  pull: always
-  image: plugins/docker
-  settings:
-    repo: cs3org/wopibridge
-    tags: ${DRONE_TAG}
-    dockerfile: wopibridge-poc.Dockerfile
-    username:
-      from_secret: dockerhub_username
-    password:
-      from_secret: dockerhub_password
-    build_args:
-      - VERSION=${DRONE_TAG}
-    custom_dns:
-      - 128.142.17.5
-      - 128.142.16.5
